+++ middleware_api/lambda/checkpoints/list_checkpoints.py
@@ -2,9 +2,9 @@ import json
 import logging
 import os
 
-from _types import CheckPoint
 from common.ddb_service.client import DynamoDbUtilsService
 from common.response import ok
+from lib._types import CheckPoint
 from multi_users._types import PARTITION_KEYS, Role
 from multi_users.utils import get_user_roles, check_user_permissions, get_permissions_by_username
