+++ middleware_api/lambda/inference/app.py
@@ -1,5 +1,5 @@
-import logging.config
 import json
+import logging.config
 import logging.config
 import os
 import traceback
